requirements.txt
langchain>=0.3.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
langchain-google-vertexai>=2.0.0,<3.0.0
langchain-google-firestore>=0.5.0,<1.0.0
google-cloud-firestore>=2.16.0,<3.0.0
google-cloud-aiplatform>=1.30.0,<2.0.0
gradio>=4.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0
